7.25 Поправлена авторизация и хедер на страницах.
