--- project/src/pages/head-guest-screen/head-guest-screen.tsx
@@ -1,25 +1,0 @@
-import Logo from '../../components/logo/logo';
-import React from 'react';
-
-function HeadGuestScreen(): JSX.Element {
-  return (
-    <section className="film-card">
-      <div className="film-card__bg">
-        <img src="img/bg-header.jpg" alt="" />
-      </div>
-
-      <h1 className="visually-hidden">WTW</h1>
-
-      <header className="page-header">
-        <Logo linkClass={'logo__link'} />
-
-        <div className="user-block">
-          <a href="sign-in.html" className="user-block__link">Sign in</a>
-        </div>
-      </header>
-
-    </section>
-  );
-}
-
-export default HeadGuestScreen;
