+++ project/src/pages/main-screen/main-screen.tsx
@@ -1,13 +1,11 @@
 import React from 'react';
-import {useAppSelector, useAppDispatch} from '../../hooks';
+import {useAppSelector} from '../../hooks';
 import FilmsList from '../../components/films-list/films-list';
 import GenresList from '../../components/genres-list/genres-list';
 import Logo from '../../components/logo/logo';
-import {logoutAction} from '../../store/api-actions';
-import {Link} from 'react-router-dom';
+import Header from '../../components/header/header';
 
 function MainScreen(): JSX.Element {
-  const dispatch = useAppDispatch();
   const promo = useAppSelector((state) => state.promo);
   const filmsList = useAppSelector((state) => state.films);
   const filteredFilmsList = useAppSelector((state) => state.filteredFilms);
@@ -20,30 +18,7 @@ function MainScreen(): JSX.Element {
         </div>
 
         <h1 className="visually-hidden">WTW</h1>
-
-        <header className="page-header film-card__head">
-          <Logo linkClass={'logo__link'} />
-
-          <ul className="user-block">
-            <li className="user-block__item">
-              <div className="user-block__avatar">
-                <img src="img/avatar.jpg" alt="User avatar" width="63" height="63"/>
-              </div>
-            </li>
-            <li className="user-block__item">
-              <Link
-                className="user-block__link"
-                onClick={(evt) => {
-                  evt.preventDefault();
-                  dispatch(logoutAction());
-                }}
-                to='/'
-              >
-                Sign out
-              </Link>
-            </li>
-          </ul>
-        </header>
+        <Header />
 
         <div className="film-card__wrap">
           <div className="film-card__info">
